fixed new message issue
